Test uploads (#52)
Possibly, premature merge, but need to add workflow file a step 0

Adding upload test for local version of ezBIDS w/ multi-modality mixed folder of dicoms
merging and pushing so that test can be triggered in actions as it's not presently in
workflows. Small changes to Makefile to ensure running of tests.
* adding local upload test to CI
* add create mixed dicom folder to makefile