Updated launch.sh to handle Mac better (#53)
Docker on Mac was having issues with mounting /tmp/, so update launch.sh
to throw a warning if it is unable to mount, create, or own a folder
mounted there. Also decreased delay in the handler in looking for new
jobs as well as disabled watch for __pycache__ in handler start with
PM2. Watch should....be disabled for local deployment and instead a
development docker compose should be written, but we're keeping this set
of changes to the minimum for now.